\name{Rwikia-package}
\alias{Rwikia-package}
\alias{Rwikia}
\docType{package}
\title{Rwikia}
\description{Wikia-specific functions and data.}
\details{
\tabular{ll}{
Package: \tab Rwikia\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-21\cr
License: \tab Proprietary\cr
LazyLoad: \tab yes\cr
}
}
\author{Jonathan March <jmarch@wikia-inc.com>}
